+++ code/model/Booking.php
@@ -1,5 +1,7 @@
 <?php
 
+use ilateral\SimpleBookings\Helpers\Syncroniser;
+
 /**
  * A single booking that can contain multiple "resources". Each resource is linked to
  * a "BookableProduct" and has a start and end date.
@@ -419,123 +421,21 @@ class Booking extends DataObject implements PermissionProvider
      */
     public function sync()
     {
-        $write = false;
 
         // If we have no order assigned, generate an estimate and
         // link to this booking
-        if (!$this->Order()->exists()) {
+        $order = $this->Order();
+        
+        if (!$order->exists()) {
             $order = Estimate::create();
             $order->write();
-        } else {
-            $order = $this->Order();
-        }
-        
-        foreach ($this->resources() as $resource) {
-            if ($resource->ProductID && $product = BookableProduct::get()->byID($resource->ProductID)) {
-                $total_time = count(
-                    SimpleBookings::create_date_range_array(
-                        $resource->Start,
-                        $resource->End,
-                        $product->PricingPeriod
-                    )
-                );
-                $item = null;
-
-                // Clean and reset any matched products
-                foreach ($order->Items() as $order_item) {
-                    $stock_item = $order_item->FindStockItem();
-
-                    if ($stock_item && $stock_item->ID == $product->ID) {
-                        $item = $order_item;
-
-                        $items_to_remove = [
-                            _t("SimpleBookings.StartDate", "Start Date"),
-                            _t("SimpleBookings.EndDate", "End Date"),
-                            _t("SimpleBookings.LengthOfTime", "Length of Time")
-                        ];
-
-                        foreach ($order_item->Customisations() as $customisation) {
-                            if (in_array($customisation->Title, $items_to_remove)) {
-                                $customisation->delete();
-                            }
-                        }
-                    }
-                }
-
-                // If we haven't found an existing order item, create a new one
-                if (!$item) {
-                    $item = OrderItem::create(
-                        array(
-                        "Key" => $product->ID,
-                        "Title" => $product->Title,
-                        "Quantity" => $resource->BookedQTY,
-                        "Price" => $product->Price,
-                        "TaxRate" => $product->TaxPercent,
-                        "StockID" => $product->StockID,
-                        "ProductClass" => $product->ClassName,
-                        "Stocked" => false,
-                        "Deliverable" => false
-                        )
-                    );
-                    $item->ParentID = $order->ID;
-                } else {
-                    $item->Quantity = $resource->BookedQTY;
-                }
-                $item->write();
-
-                // Setup customisation on an order item
-                $customisation = OrderItemCustomisation::create(
-                    array(
-                    "Title" => _t("SimpleBookings.StartDate", "Start Date"),
-                    "Value" => $product->Start,
-                    "Price" => 0
-                    )
-                );
-                $customisation->write();
-                $item->Customisations()->add($customisation);
-
-                $customisation = OrderItemCustomisation::create(
-                    array(
-                    "Title" => _t("SimpleBookings.EndDate", "End Date"),
-                    "Value" => $product->End,
-                    "Price" => 0
-                    )
-                );
-                $customisation->write();
-                $item->Customisations()->add($customisation);
-
-                $customisation = OrderItemCustomisation::create(
-                    array(
-                    "Title" => _t("SimpleBookings.LengthOfTime", "Length of Time"),
-                    "Value" => $total_time,
-                    "Price" => ($product->Price * $total_time) - $product->Price
-                    )
-                );
-                $customisation->write();
-                $item->Customisations()->add($customisation);
-            }
         }
 
-        // Ensure we sync contact details to the order
-        foreach ($this->config()->fields_to_sync as $b_field => $o_field) {
-            $order->{$o_field} = $this->{$b_field};
-        }
+        $sync = Syncroniser::create($this, $order)
+            ->setSyncProducts(true)
+            ->bookingToOrder();
 
-        $order->write();
-
-        if ($this->CustomerID != $order->CustomerID) {
-            $this->CustomerID = $order->CustomerID;
-            $write = true;
-        }
-
-        if (!$this->Order()->exists()) {
-            $this->OrderID = $order->ID;
-            $write = true;
-        }
-
-        if ($write) {
-            $this->write();
-        }
+        $this->extend("afterSync", $order, $sync);
     }
 
     /**